+++ helper/getData.js
@@ -1,0 +1,9 @@
+export const getData = async (url) => {
+  try {
+    const { data } = await axios.get(url);
+    //console.log(data);
+    return data;
+  } catch (error) {
+    console.log(error);
+  }
+};
