+++ index.html
@@ -21,89 +21,70 @@
           alt="logo"
         />
       </figure>
-      <figure class="charizard-baby">
-        <img
-          src="https://www.pngall.com/wp-content/uploads/4/Pokeball-PNG-Images.png"
-          alt="pokemon"
-        />
-      </figure>
+
       <h1>Pokédex_</h1>
-      <input type="text" class="input" id="inputbarra" placeholder="Buscar" name="search" />
-      <button class="btn btn--submit" id="btnsearch">Buscar</button>
+      <form class="form__search">
+        <input
+          id="inputSearch"
+          type="text"
+          placeholder="Enter a pokemon name"
+        />
+        <button type="submit">Buscar</button>
+      </form>
     </header>
+
     <main class="main-container" id="container">
       <section class="charizard-container">
         <div class="title">
-            <figure>
-                <img src="https://static.vecteezy.com/system/resources/previews/001/188/706/original/flame-png.png" alt="llama">
-            </figure>
-            <h2>CHARIZARD</h2>
+          <figure>
+            <img
+              class="icon"
+              src="https://static.vecteezy.com/system/resources/previews/001/188/706/original/flame-png.png"
+              alt="llama"
+            />
+          </figure>
+          <h2 class="pokemonName">CHARIZARD</h2>
         </div>
         <figure class="image-charizard">
-            <img src="https://www.pngmart.com/files/12/Charizard-PNG-Free-Download.png" alt="chaizard">
+          <img
+            class="pokemonImg"
+            src="https://www.pngmart.com/files/12/Charizard-PNG-Free-Download.png"
+            alt="chaizard"
+          />
         </figure>
-    </section>
-    <article class="info-pokemon">
+      </section>
+      <article class="info-pokemon">
         <div class="row-info">
-            <div class="row-item">
-                <span class="title-info">No.</span>
-                <span class="info-text">006</span>
-            </div>
-            <div class="row-item">
-                <span class="title-info">LEVEL</span>
-                <span class="info-text">100</span>
-            </div>
+          <div class="row-item">
+            <span class="title-info">LEVEL</span>
+            <span class="info-text level">100</span>
+          </div>
         </div>
         <div class="row-info">
-            <div class="row-item">
-                <span class="title-info">TYPE</span>
-                <span class="info-text">FIRE</span>
-            </div>
-            <div class="row-item">
-                <span class="title-info">HABILITY</span>
-                <span class="info-text">FLAMES</span>
-            </div>
+          <div class="row-item">
+            <span class="title-info">TYPE</span>
+            <span class="info-text type">FIRE</span>
+          </div>
+          <div class="row-item">
+            <span class="title-info">HABILITY</span>
+            <span class="info-text hability">FLAMES</span>
+          </div>
         </div>
         <div class="row-info">
-            <div class="row-item">
-                <span class="title-info">HEIGHT</span>
-                <span class="info-text">1.7 m</span>
-            </div>
-            <div class="row-item">
-                <span class="title-info">WEIGHT</span>
-                <span class="info-text">90.5 Kg</span>
-            </div>
+          <div class="row-item">
+            <span class="title-info">HEIGHT</span>
+            <span class="info-text height">1.7 m</span>
+          </div>
+          <div class="row-item">
+            <span class="title-info">WEIGHT</span>
+            <span class="info-text weight">90.5 Kg</span>
+          </div>
         </div>
-    </article>
+      </article>
     </main>
     <footer class="footer-content">
       <h3>OTHERS</h3>
-      <div>
-        <figure name="squirtle">
-          <img id="squirtle" name="squirtle" class="imgfooter"
-            src="https://imagenpng.com/wp-content/uploads/2017/02/007Squirtle_Pokem.png"
-            alt="pokemon" value="squirtle"
-          />
-        </figure>
-        <figure name="mewtwo">
-          <img name="mewtwo" class="imgfooter"
-            src="https://i.pinimg.com/originals/b6/db/bb/b6dbbb8add6a9042a8b5b7e972780200.png"
-            alt="pokemon"
-          />
-        </figure>
-        <figure name="charmander">
-          <img name="charmander" class="imgfooter"
-            src="https://i.pinimg.com/originals/f2/95/76/f295769d9bd3c34ffc552e837f5304ae.png"
-            alt="pokemon"
-          />
-        </figure>
-        <figure name="bulbasaur">
-          <img name="bulbasaur" class="imgfooter"
-            src="https://www.pngplay.com/wp-content/uploads/10/Bulbasaur-Pokemon-PNG-Clipart-Background.png"
-            alt="pokemon"
-          />
-        </figure>
-      </div>
+      <div id="containerCards"></div>
     </footer>
 
     <script src="./main.js" type="module"></script>
