--- sprint-1/README.md
@@ -1,12 +1,0 @@
-# Sprint 1 🐱‍💻
-## Contenido
-### Sesión 1 Versionamiento de código
-1. ¿Qué es un repositorio?
-2. Control de versiones y ¿Por qué es importante? 
-3. ¿Qué es Git?
-4. ¿Qué es GitHub?
-5. Estrategias de branching: Thunk base & Git Flow
-#### Recurso de estudio adicional
-[Guía para trabajo colaborativo en Git y GitHub](https://makaia-my.sharepoint.com/:b:/g/personal/mentoria_bootcamp_makaia_org/EYYhJgmPncdKrspBUsNxBTcB14aXL9BaXikResKRfsiC6Q?e=3MZaQd)
-## Retos, Workshops y Proyecto 🤖
-1. [Workshop 1](https://makaia-my.sharepoint.com/:b:/g/personal/mentoria_bootcamp_makaia_org/ESbSMIbewtFPkkcjW724DDsBTymvy27TL8Dyloeqc4pxtw?e=ov54Y3)
