--- sprint-1/practice-axios/index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <link rel="stylesheet" href="style.css">
-    <title>Document</title>
-</head>
-<body>
-    <header>
-        <h1>Pokemons</h1>
-        <form class="form__search">
-            <input id="inputSearch" type="text" placeholder="Enter a pokemon name">
-            <button type="submit">Buscar</button>
-        </form>
-    </header>
-    <main id="containerCards">
-        
-    </main>
-    
-    <script src="./main.js" type="module"></script>
-</body>
-</html>
